user invite / password forgotten

+++ db/migrate/20150307080000_create_organizations.rb
@@ -3,8 +3,8 @@ class CreateOrganizations < ActiveRecord::Migration
     create_table :organizations do |t|
       t.string :name
       t.string :subdomain
-      t.float :latitude
-			t.float :longitude
+      t.decimal :latitude
+			t.decimal :longitude
       t.string :place_id
       t.string :website
